elisp-tred: change tree node label when expanded/collapsed [feat]
Set up a callback to update the text label for a tree node when the
node is expanded or collapsed.

What I have in mind is to show the full lisp code in a single line
when a node is collapsed, but just show the opening paren ("(") when
the node is expanded (since the remainder of the code will be visible
by looking at the child nodes). That feature is not implemented yet --
this commit just implements the callback with some silly label
changes, to demonstrate that the dynamic label-changing works.