elisp-tred: restore syntax highlighting [bugfix]
Being able to see Emacs default syntax highlighting of the code is
hugely helpful for readability.

However, when using `push-button` for the tree's `:node` widget type,
the `push-button` widget overlays it's own `button-face` on top of the
text, which effectively hides the syntax highlighting.

For now, just use a plain `item` widget for the tree's `:node`
widget. There might be other ways to fix/restore the syntax
highlighting, but this is good enough for now.