elisp-tred: move first child into parent label [layout]
Change the tree layout a bit, so that the first child of a list
appears as part of the label for the parent/list element.

To give a concrete example, it changes the rendering of `(defun
hello () (message "hello!")) from:

[-] (
 |- defun
 |- hello
 |- ()
 |- (message "hello!"))

to:

[-] (defun
 |- hello
 |- ()
 |- (message "hello!"))

Currently, the only case where we keep the old layout is when the
first child element is a list. For example, we render `(let ((one
1) (two 2)))` as:

[-] (let
 |-[-] ( <--- note: we don't put `(one 1)` here
    |- (one 1)
    |- (two 2)))