elisp-tred: initial work on `elisp-tred--tree-mapping-rules` [wip]
Initial work to use query-based rules for mapping the structure of the
treesit parse tree to the structure of the elisp-tred tree.

This commit lays some groundwork, consisting of:

1. Adding example rules to `elisp-tred--tree-mapping-rules`.

2. Adding debugging functions to show which tree-mapping rule would be
applied to the current position in the elisp source code buffer.

The new rules aren't actually being used to build the elisp-tred tree
yet -- the old hard-coded rules are still in place.