elisp-tred: switch to `elisp-tred--tree-mapping-rules` [feat]
Switch to using `elisp-tred--tree-mapping-rules` for building the
elisp-tred tree, rather than using hard-coded rules.

Concretely, this means that:

(1) When we the label text for an expanded tree node, we use the
`:expanded-label-fn` of the matching rule from
`elisp-tred--tree-mapping-rules`.

(2) When we build the child widgets for an expanded tree node, we use
the `:child-nodes-fn` of the matching rule from
`elisp-tred--tree-mapping-rules`.

In the above two points, "the matching rule" is the first query
from `elisp-tred--tree-mapping-rules` that returns non-empty results.