elisp-tred: remove `elisp-tred-mode` (minor mode) [cleanup]
Remove the minor mode `elisp-tred-mode` and it's supporting code.

Previously, enabling the minor mode would show the elisp-tred tree for
the current file in a dedicated buffer, and disabling the minor mode
would kill the elisp-tred tree buffer.

Since then, I've implemented `elisp-tred-jump-to-tree`, which builds
and shows the elisp-tred tree for the current top-level form at POINT
(again in a dedicated buffer), without needing to explicitly enable
the minor mode. I find that usage pattern to be much smoother, so I
will just build on that going forward.