elisp-tred: change `:capture-fn` -> `:match-fn` [refactor][simplify]
Rename `:capture-fn` to `:match-fn` and change the behaviour so that
it just returns nil or non-nil, to indicate if the tree-mapping
rule is a match for a given treesit node (`node`).

Previously, the `:capture-fn` was both testing for a match against
`node` and also determining `captures`, i.e. the treesit nodes to pass to
`:expander-label-fn` and `:child-nodes-fn`. But
I realized that the capturing part was unnecessary, and it was a lot
simpler to only pass `node` to `:expander-label-fn`/`:child-nodes-fn`
and let those functions figure out the relevant treesit nodes themselves.

The whole capturing idea was just a holdover from when I was trying to
use tree-sitter "query captures" to implement the tree-mapping
rules. (I found that the query-based approach was way more
confusing/error-prone and it turned out to be more verbose too, which
was the opposite of what I expected.)