elisp-tred: replace tree icon with left bracket ("(" or "[") [style]
Replace the tree-widget.el's default green/red diamond-shaped icons
with the opening bracket ("(" or "[") for the lisp form.

The tree icons are the things that the user clicks on to expand or
collapse a tree node (i.e. show or hide the node's children).

I discovered that using the opening bracket instead of
tree-widget.el's is more compact and removes unnecessary visual noise.