elisp-tred: move `:treesit-node` to `:node` widget  [refactor]
In the (uncommitted) code that's I'm currently working on, I ran into
the problem that I couldn't get the treesit node corresponding to a
leaf node, because I was only adding the `:treesit-node` property to
internal (non-leaf) tree nodes.

This commit fixes that, by moving the `:treesit-node` property to the
tree node label widgets instead (i.e. the `:node` widgets).