elisp-tred: hide tree icon/button widgets [UX]
Hide/disable the tree icon widgets, by replacing them with zero-width
widgets (i.e. set `:tag` to empty string "").

Up until now, I was using a hacky approach where the initial left
bracket ("(" or "[") of a list/vector was displayed in the tree
icon/button, and the rest of the list/vector was displayed in the tree
node label.

In this commit, I have reworked things to completely hide the
icon/button widgets, and to show the full elisp list/vector in the
tree node label (when collapsed). That approach is way more intuitive,
and it should hopefully save me some confusion in the future.