elisp-tred: change tree guides to Unicode box-drawing chars [style]
For the tree guides/handles, use Unicode box-drawing characters
instead of tree-widget.el's default image files.

The default guide/handle PNG/XPM images provided by tree-widget.el are
quite nice, but rendering the tree as pure text has a few advantages:

* I can easily increase the indent size if I want to (i.e. the length
of the handles)

* I can easily add a space between the handles and the labels,
to improve readability (as I have done in this commit)

* The rendering of the guides/handles doesn't break when I increase
the text size in the buffer

* I can easily copy/paste real tree renderings into code comments and
other documentation

In addition to the above points, I like the readability and aesthetics
of the new version a little better (probably because the handle
lengths are longer).