elisp-tred: `elisp-tred-jump-to-tree` auto-positions cursor [UX]
When opening the elisp-tred buffer, automatically move the cursor to
the tree node that corresponds to the current location in the elisp
source code buffer, expanding ancestor tree nodes as needed to expose
the target node. Even better, move the cursor to the exact offset
within the tree node label text that corresponds to elisp source code
position.