elisp-tred: smarter collapsed node labels [empty]
Rather than sequentially concatenating balanced forms until we hit the
label length limit, build up the label in a breadth-first manner, which
gives a much better overview of the code contained in the collapsed
subtree.

For example, consider the following (silly) elisp code:

    (let ((msg "hello")) (message msg))

Suppose our maximum label length is 30 characters.

With the current algorithm, the truncated label is:

    "(let ((msg "hello"))...)"

This is pretty intuitive. We just concatentate as many balance forms as we
fit within the length, and then indicate the remainder with an
ellipsis.

However, with the new algorithm, we iterative expand each
successive level in the tree until we hit the label length
limit. This results in:

    "(let ((msg...)) (message msg))"