elisp-tred: add tree navigation commands [feat][wip]
Implement some commands for navigating the tree structure
(e.g. `elisp-tred-goto-parent`, `elisp-tred-goto-children`), and add
default keybindings for them in `elisp-tred-mode-map`.

These new commands are *mostly* working, but there are still some
bugs. For example, `elisp-tred-goto-parent` doesn't work when
positioned on the child of a quoted list. My thoughts are currently a
bit confused about to how to properly/reliably implement these
navigation commands, given that the structure of the elisp-tred tree
is often not the same as the structure of the tree-sitter parse tree.