elisp-tred: rewrite nav commands to use `tree-widget' structures [refactor]
My original implementation of the navigation commands (e.g.
`elisp-tred-goto-next-sibling`) performed the navigation by querying
the structure of the tree-sitter tree, and then searching through the
buffer for the target treesit node using
`elisp-tred--treesit-node-for-current-line`.  However, this new
version of the navigation commands instead queries the
`:parent`/`:children` properties of the `tree-widget.el` data
structure for the target node / buffer location

The new implementation makes more sense because the structure of
elisp-tred tree is often note the same as the structure of the
tree-sitter tree. We often choose to make the elisp-tred tree
structure different than the tree-sitter parse tree for the sake of
readability and compactness.  For example, for lists, we combine the
open paren (`(`) and the first element of the list into a
single (parent) node, unless the first element is not itself a
list. Similarly, we combine the quote and the quoted symbol/list into
a single tree node.

The original navigation commands would work most of the time, but
would fail to move the cursor to the desired location in cases where
the elisp-tred tree structure differed from the tree-sitter parse
tree.