elisp-tred: fix hanging when `M-.` is on C function [bugfix]
Fix Emacs hanging when attempting to jump to C function definitions
with `M-.` (`xref-find-definitions`).

Elisp-tred was trying to parse the C code as if were elisp, and making
Emacs' spin indefinitely as a result.

This commit adds a fix that checks if the xref is a C function or an
elisp function before jumping to the definition. In the case that the
xref is a C function, we fall back to the default behaviour of jumping
to the C source file, rather than trying to construct an elisp-tred
buffer/tree.