elisp-tred: add `:collapsed-label-fn` [refactor]
Make the collapsed label for a elisp-tred node customizable, but
adding a `:collapsed-label-fn` property to
`elisp-tred--tree-mapping-rules`.

I initially thought I wouldn't need a `:collapsed-label-fn` property,
because the collapsed labels would always be generated in the same way
for all nodes, i.e. the full elisp source code text with newlines and
duplicated spaces removed. But so far, I've ended up implementing
different logic for lisp atoms and lisp sequences (lists or vectors),
and I'm planning to implement special logic for strings too (so that
multi-line docstrings are more readable.) So my initial assumption
about not needing to customize collapsed labels seems to be wrong.