elisp-tred: fix syntax-highlighting on jump-to-definition [bugfix]
Force syntax-highlighting of the elisp code before creating the
corresponding node label in the elisp-tred tree.

By default, Emacs does lazy/just-in-time syntax-highlighting of elisp
code, where the code does not get syntax-highlighted until it becomes
visible (i.e. the user scrolls to it in a window).

As a result, an elisp-tred tree that was rendered as the result of
calling `M-.` (`xref-find-definitions`) on a symbol in an elisp-tred
buffer would usually not have syntax-highlighting.  The only time when
the new tree would be correctly syntax-highlighted was when the user
already happened to view the target code in an `emacs-lisp-mode`
buffer.