elisp-tred: render tree for entire file [UX]
When calling `elisp-tred`, include the code for the entire file in the
tree, not just the code for the current function.

IMO, showing the code for entire file is a more intuitive default,
because programmers are long accustomed to organizing and editing
source code at the file level.

We can always implement narrowing/widening functions to restore the
old function-centric behaviour later, if desired.